Class Conversion initialized.
I began to create Conversion class, which will be an abstract class serving as
a template for all the other conversion classes (Binary, Hexadecimal ...)

For the moment the class is not abstract for experimental reasons. But it'll be pretty soon.

A ASCII conversion method has been implemented. Converting each caracters to ascii and prints the whole.